1. moved the corpus statistics under the text statistics GUI and out of style analysis.
