+++ lib/release_version.txt
@@ -1,1 +1,1 @@
-4.4.1
+4.4.2
