+++ src/statistics_txt_main.py
@@ -4,41 +4,91 @@ import sys
 import GUI_util
 import IO_libraries_util
 
-if IO_libraries_util.install_all_Python_packages(GUI_util.window,"statistics_txt_ALL_main.py",['os','csv','tkinter','ntpath','collections','subprocess'])==False:
+if IO_libraries_util.install_all_Python_packages(GUI_util.window,"statistics_txt_main.py",['os','csv','tkinter','ntpath','collections','subprocess'])==False:
     sys.exit(0)
 
 import os
+import tkinter.messagebox as mb
 import tkinter as tk
 from subprocess import call
 
 import GUI_IO_util
 
+corpus_statistics_var = tk.IntVar()
+corpus_statistics_options_menu_var = tk.StringVar()
+corpus_text_options_menu_var = tk.StringVar()
+
 # RUN section ______________________________________________________________________________________________________________________________________________________
 
-def run(inputFilename, inputDir, outputDir,openOutputFiles,createCharts,chartPackage,
-    n_grams_var,
-    corpus_statistics_var):
+def run(inputFilename, inputDir, outputDir,
+        corpus_statistics_options_menu_var,
+        corpus_text_options_menu_var,
+        openOutputFiles,createCharts,chartPackage,
+        corpus_statistics_var):
 
     filesToOpen = []  # Store all files that are to be opened once finished
-    openOutputFilesSV=openOutputFiles
-    outputDir_style=outputDir
 
-    if n_grams_var == True:
-        call("python NGrams_CoOccurrences_main.py", shell=True)
+    if not corpus_statistics_var:
+        mb.showwarning('Warning',
+                       'No option has been selected.\n\nPlease, select an option and try again.')
         return
+
+    # corpus statistics --------------------------------------------------------------------
 
     if corpus_statistics_var:
-        call("python style_analysis_main.py", shell=True)
-        return
+        stopwords_var = False
+        lemmatize_var = False
+        if corpus_text_options_menu_var == '*':
+            stopwords_var = True
+            lemmatize_var = True
+        if 'Lemmatize' in corpus_text_options_menu_var:
+            lemmatize_var = True
+        if 'stopwords' in corpus_text_options_menu_var:
+            stopwords_var = True
+        if "*" in corpus_statistics_options_menu_var or 'frequencies' in corpus_statistics_options_menu_var:
+
+            import statistics_txt_util
+            outputFiles = statistics_txt_util.compute_corpus_statistics(window, inputFilename, inputDir, outputDir,
+                                                                        config_filename, False, createCharts,
+                                                                        chartPackage, stopwords_var, lemmatize_var)
+            if outputFiles != None:
+                if isinstance(outputFiles, str):
+                    filesToOpen.append(outputFiles)
+                else:
+                    filesToOpen.extend(outputFiles)
+        if "Compute sentence length" in corpus_statistics_options_menu_var or "*" in corpus_statistics_options_menu_var:
+            outputFiles = statistics_txt_util.compute_sentence_length(inputFilename, inputDir, outputDir,
+                                                                      config_filename, createCharts, chartPackage)
+            if outputFiles != None:
+                if isinstance(outputFiles, str):
+                    filesToOpen.append(outputFiles)
+                else:
+                    filesToOpen.extend(outputFiles)
+
+        if "Compute line length" in corpus_statistics_options_menu_var or "*" in corpus_statistics_options_menu_var:
+            outputFiles = statistics_txt_util.compute_line_length(window, config_filename, inputFilename, inputDir,
+                                                                  outputDir,
+                                                                  False, createCharts, chartPackage)
+            if outputFiles != None:
+                if isinstance(outputFiles, str):
+                    filesToOpen.append(outputFiles)
+                else:
+                    filesToOpen.extend(outputFiles)
+
+    if openOutputFiles == True:
+        import IO_files_util
+        IO_files_util.OpenOutputFiles(GUI_util.window, openOutputFiles, filesToOpen, outputDir, scriptName)
+
 
 #the values of the GUI widgets MUST be entered in the command otherwise they will not be updated
 run_script_command=lambda: run(GUI_util.inputFilename.get(),
                                 GUI_util.input_main_dir_path.get(),
                                 GUI_util.output_dir_path.get(),
+                                corpus_statistics_options_menu_var.get(),
+                                corpus_text_options_menu_var.get(),
                                 GUI_util.open_csv_output_checkbox.get(),
                                 GUI_util.create_chart_output_checkbox.get(),
                                 GUI_util.charts_package_options_widget.get(),
-                                n_grams_var.get(),
                                 corpus_statistics_var.get())
 
 GUI_util.run_button.configure(command=run_script_command)
@@ -83,20 +133,16 @@ inputFilename=GUI_util.inputFilename
 GUI_util.GUI_top(config_input_output_numeric_options, config_filename, IO_setup_display_brief, scriptName)
 
 def clear(e):
-    n_grams_checkbox.configure(state='normal')
     corpus_statistics_checkbox.configure(state='normal')
 
-    n_grams_var.set(0)
     corpus_statistics_var.set(0)
+    corpus_statistics_options_menu_var.set('')
+    corpus_text_options_menu_var.set('')
 
     GUI_util.clear("Escape")
 window.bind("<Escape>", clear)
 
 # GUI CHANGES cut/paste special GUI widgets from GUI_util
-
-n_grams_list=[]
-
-bySentenceIndex_var=tk.IntVar()
 
 n_grams_button = tk.Button(window, width=GUI_IO_util.widget_width_short, text='Compute N-Grams (Open GUI)',  command=lambda: call("python style_analysis_main.py", shell=True))
 # place widget with hover-over info
@@ -105,12 +151,41 @@ y_multiplier_integer = GUI_IO_util.placeWidget(window,GUI_IO_util.labels_x_coord
                                    False, False, True, False, 90, GUI_IO_util.labels_x_coordinate,
                                    "Click on the button to open the GUI")
 
-corpus_statistics_button = tk.Button(window, width=GUI_IO_util.widget_width_short, text='Compute document(s) statistics (Open GUI)',  command=lambda: call("python style_analysis_main.py", shell=True))
+
+corpus_statistics_var.set(0)
+corpus_statistics_checkbox = tk.Checkbutton(window,text="Compute document(s) statistics", variable=corpus_statistics_var, onvalue=1, offvalue=0)
+y_multiplier_integer=GUI_IO_util.placeWidget(window,GUI_IO_util.labels_x_coordinate,y_multiplier_integer,corpus_statistics_checkbox,True)
+
+corpus_statistics_options_menu_var.set('*')
+
+corpus_statistics_options_menu = tk.OptionMenu(window,corpus_statistics_options_menu_var,
+                                                '*',
+                                               'Compute frequencies of sentences, words, syllables, and top-20 words',
+                                               'Compute sentence length',
+                                               'Compute line length',
+                                               )
 # place widget with hover-over info
-y_multiplier_integer = GUI_IO_util.placeWidget(window,GUI_IO_util.labels_x_coordinate, y_multiplier_integer,
-                                   corpus_statistics_button,
-                                   False, False, True, False, 90, GUI_IO_util.labels_x_coordinate,
-                                   "Click on the button to open the GUI")
+y_multiplier_integer = GUI_IO_util.placeWidget(window,GUI_IO_util.IO_configuration_menu, y_multiplier_integer,
+                                   corpus_statistics_options_menu,
+                                   True, False, True, False, 90, GUI_IO_util.IO_configuration_menu,
+                                   "Select the statistics option for your document(s) (* for all); widget disabled until checkbox ticked.")
+
+# y_multiplier_integer=GUI_IO_util.placeWidget(window,GUI_IO_util.style_corpus_statistics_options_menu_pos,y_multiplier_integer,corpus_statistics_options_menu, True)
+
+corpus_text_options_menu_var.set('')
+corpus_options_menu_lb = tk.Label(window, text='Text options')
+y_multiplier_integer=GUI_IO_util.placeWidget(window,GUI_IO_util.style_corpus_text_options_menu_lb_pos,y_multiplier_integer,corpus_options_menu_lb,True)
+corpus_text_options_menu = tk.OptionMenu(window, corpus_text_options_menu_var, '*','Lemmatize words', 'Exclude stopwords & punctuation')
+y_multiplier_integer=GUI_IO_util.placeWidget(window,GUI_IO_util.style_corpus_options_menu_men_pos,y_multiplier_integer,corpus_text_options_menu)
+
+def activate_corpus_options(*args):
+    if corpus_statistics_var.get()==True:
+        corpus_statistics_options_menu.configure(state='normal')
+        corpus_text_options_menu.configure(state='normal')
+    else:
+        corpus_statistics_options_menu.configure(state='disabled')
+        corpus_text_options_menu.configure(state='disabled')
+corpus_statistics_var.trace('w',activate_corpus_options)
 
 videos_lookup = {'No videos available':''}
 videos_options='No videos available'
@@ -140,7 +215,7 @@ def help_buttons(window,help_button_x_coordinate,y_multiplier_integer):
 
     y_multiplier_integer = GUI_IO_util.place_help_button(window,help_button_x_coordinate,y_multiplier_integer,"NLP Suite Help",'Please, tick the \'N-Grams\' checkbox if you wish to open the N-Grams GUI to analyze n-grams (1, 2, 3) present in your corpus.')
     y_multiplier_integer = GUI_IO_util.place_help_button(window, help_button_x_coordinate, y_multiplier_integer, "NLP Suite Help",
-                                  'Please, tick the checkbox if you wish to open the style GUI to compute basic statistics on your corpus.')
+                                'Please, tick the checkbox if you wish to compute basic statistics on your corpus.\n\nUse the dropdown menu to select the desired option. Use the Text options dropdown menu to lemmatize words or exclude stopwords/punctuation.\n\nIn INPUT the script expects a single txt file or a directory containing a set of txt files.\n\nIn OUTPUT, the script generates a set of csv files each containing word n-grams between 1 and 3.\n\nWhen n-grams are computed by sentence index, the sentence displayed in output is always the first occurring sentence.')
     y_multiplier_integer = GUI_IO_util.place_help_button(window,help_button_x_coordinate,y_multiplier_integer,"NLP Suite Help",GUI_IO_util.msg_openOutputFiles)
     return y_multiplier_integer -1
 y_multiplier_integer = help_buttons(window,GUI_IO_util.help_button_x_coordinate,0)
